refactor(ai): split tools.rs god-file into per-domain dispatch modules (#345)
Bump version to 0.4.100. ai/tools.rs was 1446 lines, over the 1000-line cap.
Decompose into ai/tools/ submodules with execute_tool() as a thin dispatcher:

  mod.rs              — shared helpers + execute_tool dispatch + tool_definitions re-export
  library.rs          — library + (non-bible) presentation tools
  slides.rs           — presentation-slide editing + stage triggers
  bible.rs            — bible lookup / verse loading / slide resolution / trigger
  bible_presentation.rs — bible-presentation CRUD (server-side composition)
  misc.rs             — get_style_guide
  tests.rs            — existing test module (moved verbatim)

Pure structural refactor — every tool branch moved verbatim into a handler
function; no behavior change. Public path tools::execute_tool / tools::tool_definitions
preserved (agent.rs untouched). All 245 presenter-server tests pass (incl. the
31 tool tests), clippy clean, fmt clean. Largest module now 268 lines.

Note: misc.rs uses include_str!("../style_guide.md") since the file moved one
directory deeper.

Closes #345
